Remove migrations directory
The migrations/ directory is auto-generated by manage.py. If the
directory exists, and the command `python manage.py db init` is run,
the following error is returned:

Error: Directory migrations already exists

A quick fix for this error is simply to delete the directory, and remove
it from version control.

The migrations/ directory has been added to the .gitignore. It should
be removed from Git's cache with

git rm -r --cached .

https://www.git-scm.com/docs/git-rm